added input system

+++ src/events/new_game.rs
@@ -1,5 +1,5 @@
 use bevy::prelude::*;
 
 pub struct NewGameEvent {
-    pub map_size:u32
+    pub map_size:usize
 }
